+++ client/views/comments/comment_item.js
@@ -152,8 +152,8 @@ Template[getTemplate('comment_item')].events({
   'click .not-upvoted .upvote': function(e, instance){
     e.preventDefault();
     if(!Meteor.user()){
-      Router.go(getSigninUrl());
-      throwError(i18n.t("please_log_in_first"));
+      Router.go('atSignIn');
+      flashMessage(i18n.t("please_log_in_first"), "info");
     }
     Meteor.call('upvoteComment', this, function(error, result){
       trackEvent("post upvoted", {'commentId':instance.data._id, 'postId': instance.data.post, 'authorId':instance.data.userId});
@@ -162,8 +162,8 @@ Template[getTemplate('comment_item')].events({
   'click .upvoted .upvote': function(e, instance){
     e.preventDefault();
     if(!Meteor.user()){
-      Router.go(getSigninUrl());
-      throwError(i18n.t("please_log_in_first"));
+      Router.go('atSignIn');
+      flashMessage(i18n.t("please_log_in_first"), "info");
     }
     Meteor.call('cancelUpvoteComment', this, function(error, result){
       trackEvent("post upvote cancelled", {'commentId':instance.data._id, 'postId': instance.data.post, 'authorId':instance.data.userId});
@@ -172,8 +172,8 @@ Template[getTemplate('comment_item')].events({
   'click .not-downvoted .downvote': function(e, instance){
     e.preventDefault();
     if(!Meteor.user()){
-      Router.go(getSigninUrl());
-      throwError(i18n.t("please_log_in_first"));
+      Router.go('atSignIn');
+      flashMessage(i18n.t("please_log_in_first"), "info");
     }
     Meteor.call('downvoteComment', this, function(error, result){
       trackEvent("post downvoted", {'commentId':instance.data._id, 'postId': instance.data.post, 'authorId':instance.data.userId});
@@ -182,8 +182,8 @@ Template[getTemplate('comment_item')].events({
   'click .downvoted .downvote': function(e, instance){
     e.preventDefault();
     if(!Meteor.user()){
-      Router.go(getSigninUrl());
-      throwError(i18n.t("please_log_in_first"));
+      Router.go('atSignIn');
+      flashMessage(i18n.t("please_log_in_first"), "info");
     }
     Meteor.call('cancelDownvoteComment', this, function(error, result){
       trackEvent("post downvote cancelled", {'commentId':instance.data._id, 'postId': instance.data.post, 'authorId':instance.data.userId});