+++ packages/telescope-blank/package.js
@@ -15,22 +15,20 @@ Package.onUse(function (api) {
   // automatic (let the package specify where it's needed)
 
   api.use([
-    'telescope-base',
-    'telescope-lib',
-    'telescope-i18n',
-    'tap:i18n'
+    'tap:i18n',                   // internationalization package
+    'iron:router',                // routing package
+    'telescope-base',             // basic Telescope hooks and objects
+    'telescope-lib',              // useful functions
+    'telescope-i18n',             // internationalization wrapper
+    'fourseven:scss'              // SCSS compilation package
   ]);
-
-  // both
-
-  api.use([
-    //
-  ], ['client','server']);
 
   // client
 
   api.use([
-    //
+    'jquery',                     // useful for DOM interactions
+    'underscore',                 // JavaScript swiss army knife library
+    'templating'                  // required for client-side templates
   ], ['client']);
 
   // server
@@ -50,19 +48,22 @@ Package.onUse(function (api) {
   // both
 
   api.add_files([
-    'lib/both.js'
+    'lib/both.js',
+    'lib/routes.js'
   ], ['client', 'server']);
 
   // client
 
   api.add_files([
-    'lib/client/client_only.js'
+    'lib/client/templates/template.html',
+    'lib/client/templates/template.js',
+    'lib/client/scss/package.scss'
   ], ['client']);
 
   // server
 
   api.add_files([
-    'lib/server/server_only.js'
+    'lib/server/publications.js'
   ], ['server']);    
 
   // i18n languages (must come last)
@@ -79,7 +80,7 @@ Package.onUse(function (api) {
   // -------------------------------- 3. Variables to export --------------------------------
 
   api.export([
-    //
+    'myFunction'
   ]);
 
 });