+++ History.md
@@ -1,3 +1,26 @@
+## v0.10.0 “RefactorScope”
+
+* Renaming Errors to Messages (thanks @yourcelf!).
+* Adding `fast-render` (thanks @arunoda!).
+* Extracted digest into its own package.
+* Adding "scheduled" view that shows upcoming scheduled posts.
+* Bringing back "delete post" link that was removed by mistake.
+* Made user profile display page modular.
+* Made user profile edit page modular.
+* Started extracting invites into their own package.
+
+## v0.9.11 “FormScope”
+
+* Now using [Autoform](https://github.com/aldeed/meteor-autoform/)'s **quickform** feature to generate post submit and edit forms. 
+* Various fixes by [@anthonymayer](https://github.com/anthonymayer).
+* Now using [fourseven:scss](https://github.com/fourseven/meteor-scss) to directly compile SCSS files. 
+* Renamed `post` method to `submitPost`. 
+* Post editing now happens via a `postEdit` method.
+* Categories are now normalized (only the `_id` is stored on the post object, not the whole category object).
+* Refactored Embedly package; now fills in description as well (thanks [@kvindasAB](https://github.com/kvindasAB)!).
+* Thumbnail height and width are now customizable in settings panel.
+* Settings and Post forms now i18n'ized.
+
 ## v0.9.10 “i18nScope”
 
 * Now using [tap:i18n](https://github.com/TAPevents/tap-i18n) for internationalization (thanks a ton to @theosp).